[#30] refactor: 인증관련 로직 auth 패키지로 이동
