+++ api/src/main/java/com/grayzone/domain/user/controller/UserController.java
@@ -1,19 +1,16 @@
 package com.grayzone.domain.user.controller;
 
 import com.grayzone.common.ResponseDataDto;
-import com.grayzone.domain.user.UserTerm;
-import com.grayzone.domain.user.dto.request.SignUpRequestDto;
 import com.grayzone.domain.user.dto.request.VerifyNicknameDuplicateRequestDto;
-import com.grayzone.domain.user.dto.response.SignUpResponseDto;
-import com.grayzone.domain.user.dto.response.TermsResponseDto;
 import com.grayzone.domain.user.service.UserService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.stream.Stream;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/api")
@@ -31,28 +28,5 @@ public class UserController {
     return ResponseEntity
       .status(HttpStatus.NO_CONTENT)
       .body(ResponseDataDto.from(null));
-  }
-
-  @GetMapping("/terms")
-  public ResponseEntity<ResponseDataDto<TermsResponseDto>> getTerms() {
-    return ResponseEntity.ok(
-      ResponseDataDto.from(
-        new TermsResponseDto(
-          Stream.of(UserTerm.values()).map(term ->
-            new TermsResponseDto.TermResponseDto(term.isRequired(), term.getTitle(), term.getUrl(), term.getCode())
-          ).toList())
-      )
-    );
-  }
-
-  @PostMapping("/users")
-  public ResponseEntity<ResponseDataDto<SignUpResponseDto>> signUp(
-    @RequestBody SignUpRequestDto requestDto
-  ) {
-    return ResponseEntity.ok(
-      ResponseDataDto.from(
-        userService.signUp(requestDto)
-      )
-    );
   }
 }
